// src/types/interfaces/monday-item-creation-columns-types.ts
export interface MondayItemCreationColumnsTypes {
  clientNameColumn: string;
  clientIdColumn: string;
  businessStreamNameColumn: string | undefined;
  businessStreamIdColumn: string;
  legalEntityNameColumn: string | undefined;
  legalEntityCountryColumn: string | undefined;
  obligationDefinitionTypeTextColumn: string | undefined;
  obligationDefinitionTypeIdColumn: string | undefined;
  taxRegistrationColumn: string | undefined;
  complexityColumn: string;
  commentsColumn: string | undefined;
}